Revert "New Inference benchmarks (#382)"
This reverts commit c44127eadf38058fcd33e44f8366cf8159b7f5a4.

+++ benchmarks/brax/voirfile.py
@@ -22,7 +22,7 @@ class Config:
     stop: int = 60
 
     # Number of seconds between each gpu poll
-    gpu_poll: float = 1
+    gpu_poll: int = 1
 
 
 @configurable
